Added tests for Dashboard Actions
