+++ src/actions/Login.ts
@@ -16,6 +16,7 @@ if(!window.localStorage)
 } else{
     localStorage=window.localStorage;
 }
+
 export enum ActionTypes {
     CHECK_OSIO_LOGIN = '[login] CHECK_OSIO_LOGIN',
     CHECK_CHE_LOGIN = '[login] CHECK_CHE_LOGIN',
